backend/requirements.txt
fastapi
uvicorn
python-dotenv
transformers
torch
pydantic